More options for TPV modifications (lwrad sees modified stratospheric qv and spatial extent is limited by fixed contour value).
A test where the lwrad sees qv(strato)=.1*qvOriginal had a dipole of a
pv anomaly with a large + anomaly above the tropopause and a small -
anomaly below, consistent with increased cooling at trop level. It may
be better to use a smoother qv profile than the jump we introduced. The amplitude of the TPV had a marked linear
intensification with time.